[#3840] Extract shared OTP machinery into MfaFlowHelper
Both full_mfa/ specs duplicated the same ~120 lines verbatim:
ENV set, Rack::Test, app, identities, provision_totp, allow_immediate_otp_reuse!,
json_post, json_body, clear_body_headers, seed_account_with_password.

Move it all into support/mfa_flow_helper.rb (mirroring oauth_flow_helper.rb)
and replace each copy with a single require + include MfaFlowHelper. Route
driving stays in each spec — that is the part that genuinely differs.
